changed OrderHistory to DonationHistory and updated accordingly. Added thought bubble images, and imported to homepage, and added styling
