+++ client/src/App.css
@@ -40,4 +40,3 @@
 .read-the-docs {
   color: #888;
 }
-
