+++ client/src/components/Nav/index.jsx
@@ -9,7 +9,7 @@ function Nav() {
       return (
         <ul className="flex-row">
           <li className="mx-1">
-            <Link to="/orderHistory">
+            <Link to="/donationHistory">
               Make a Donation
             </Link>
           </li>
