+++ client/src/index.css
@@ -15,14 +15,9 @@
   box-sizing: border-box;
 }
 
-.homeContainer{
-  background-image: linear-gradient(to top left, #c490d1, #acacde, #abdafc);
-}
-
 html,
 body {
-  min-height: 100%;
-
+  min-height: 100%;  
 }
 
 body {
@@ -215,3 +210,4 @@ input:focus {
     width: 100%;
   }
 }
+
