+++ client/src/pages/DonationHistory.jsx
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 import { useQuery } from '@apollo/client';
 import { QUERY_USER } from '../utils/queries';
 
-function OrderHistory() {
+function DonationHistory() {
   const { data } = useQuery(QUERY_USER);
   let user;
 
@@ -48,4 +48,4 @@ function OrderHistory() {
   );
 }
 
-export default OrderHistory;
+export default DonationHistory;
